security+integrity: full-stack zero-trust audit phase 2 — IDOR defense, GL atomicity, decimal fix, CSRF, CSS
BACKEND SECURITY:
- routes/ai.py: replace db.get_or_404(Product) with get_owned_or_404
- routes/whatsapp.py: replace db.get_or_404(Sale, Customer) with get_owned_or_404
- routes/reports.py: replace db.get_or_404(Supplier, Customer) with get_owned_or_404
- routes/ledger.py: replace db.get_or_404(GLJournalEntry) with get_owned_or_404
- routes/graphql.py: block all GraphQL mutations (read-only layer); prevent cross-tenant sale creation
- services/graphql_service.py: replace db.session.get in resolvers with get_owned_or_404
- services/cheque_accounting_integration.py: replace all db.get_or_404(Cheque) with get_owned_or_404
- services/return_service.py: replace db.session.get(Sale,SaleLine,Product) with get_owned_or_404
- services/stock_service.py: replace db.session.get(Product) with get_owned_or_raise
- services/payment_service.py: replace db.session.get(Customer,Sale) with get_owned_or_raise
- services/ai_service.py: replace db.session.get(Product,Customer) with get_owned_or_raise
- services/sale_service.py: replace db.session.get(Warehouse,Customer) with get_owned_or_raise
- services/gl_service.py: replace db.session.get(GLJournalEntry) in reverse_entry with get_owned_or_raise
- services/advanced_journal_manager.py: wrap update/reverse/delete/approve in try/except+rollback; validate all lines BEFORE delete; use get_owned_or_404
- utils/decorators.py: add assert_same_tenant + get_owned_or_raise for non-aborting service-layer checks
- gl_service.py: replace _unique_entry_number in-process counter with generate_number (Redis distributed lock + retry) for atomic JE numbering

DATABASE & MODELS:
- models/package.py: price Float->Numeric(15,3); amount_paid Float->Numeric(15,3)
- models/expense.py: currency default 'AED'->'ILS' (system consistency)
- migrations/12_decimal_indexes_currency_fix.py: Alembic migration for Decimal conversion, 8 missing FK indexes, currency default update

FRONTEND SECURITY:
- 9 templates (hr/*, erp/create_lot, approvals, admin/ledger): add explicit CSRF token inputs to all POST forms
- static/css/accessibility.css: fix !important typo (Chinese -> correct keyword)

TESTS: 63/63 critical security & isolation tests passing

+++ models/package.py
@@ -14,7 +14,7 @@ class Package(db.Model):
     name_en = db.Column(db.String(100), nullable=False)  # اسم الباقة بالإنجليزية
     slug = db.Column(db.String(50), unique=True, nullable=False)  # للرابط
     icon = db.Column(db.String(50), default='📦')  # أيقونة الباقة
-    price = db.Column(db.Float, nullable=False)  # السعر بالدولار
+    price = db.Column(db.Numeric(15, 3), nullable=False)  # السعر بالدولار
     currency = db.Column(db.String(10), default='USD')
 
     # وصف الباقة
@@ -103,7 +103,7 @@ class PackagePurchase(db.Model):
     # بيانات الدفع
     payment_method = db.Column(db.String(50), nullable=False)  # crypto, card, paypal, bank
     payment_status = db.Column(db.String(50), default='pending')  # pending, completed, failed, refunded
-    amount_paid = db.Column(db.Float, nullable=False)
+    amount_paid = db.Column(db.Numeric(15, 3), nullable=False)
     currency = db.Column(db.String(10), default='USD')
 
     # معلومات إضافية للدفع
